RF: Create abstract FileBasedImage / FileBasedHeader classes.
